+++ src/reader.rs
@@ -1,4 +1,19 @@
-use crate::*;
+use crate::{
+    pcl_utils::*,
+    Point,
+    PointCloud2Msg,
+    PointConvertible,
+    ConversionError,
+    MetaNames,
+    PointMeta,
+    convert::{
+        FromBytes,
+        FieldDatatype,
+        load_loadable,
+        Endianness,
+        check_coord,
+    },
+};
 
 /// Convenience type for a Reader that reads coordinates as f32. Specify the number of dimensions, metadata dimensions and C, the point type.
 pub type ReaderF32<const DIM: usize, const METADIM: usize, C> =
